Friends Reload when you log back in
Since we implemented the googleMap as a fragment instead of a view, it
has to be implemented in MainActivity (or some Activity). At least, this
is the impression I got from reading the google maps reference docs.
That meant I had to keep the mapFragment where it is (in MainActivity),
and let mainFragment know about mainActivity so that it could modify the
needsToLoadFriends variable, and then call the loadFriendsData()
functions.

This implementation feels wrong to me. I'm not sure mainFragment should
have the ability to access any of mainActivity's member functions but I
can't articulate a reason why not. Anyway, it works for now, so we can
discuss it later if it's an issue.